requirements.txt

streamlit
pypdf
langchain
python-dotenv
faiss-cpu #矢量存储
openai
huggingface-hub

InstructorEmbedding
sentence_transformers